feat!: updating ATLAS color palette to match collboration recommendations (#574)
* Updating ATLAS color palette

After an internal working group to discuss color palette options and
color-vision deficiency of various kinds, these are the proposed colors
that the experiment will adopt.

* Attempting updates following review

Giving a few things a try here:
- Breaking out ATLAS style into ATLAS1 and ATLAS2
- Copying styles around a bit
- Keeping the old (disfavored) color scheme
- Linking to documentation for ATLAS members
- Updating the mplhep documentation itself

I hope I got all this right, but further comment is always welcome.

* Adding quotes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Imports

* Updating documentation and colors again in advance of the presentation

It appears that the working group has re-converged, but there will be an
additional discussion this week on the palettes, and probably we should
resist the temptation to merge this until then. Review is of course
welcome nevertheless.

* An additional style without the read

* Updating ATLAS color palettes with final recommendations

We believe these are now the final recommendations of the collaboration.

* Spelling mistake

* Spelling mistake

* Adding ATLAS1 and ATLAS2 explicitly to the styles to be built for the gallery

* Adding links to documentation

* Another place where the styles need to be listed...

---------

Co-authored-by: Zach Marshall <ZLMarshall@lbl.gov>
Co-authored-by: Zach Marshall <zmarshal@lxplus994.cern.ch>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Andrzej Novak <novak5andrzej@gmail.com>